+++ src/hyp3_isce2/insar_tops_burst.py
@@ -1,13 +1,15 @@
-"""Create a single-burst Sentinel-1 geocoded unwrapped interferogram using ISCE2's TOPS processing workflow"""
+"""Create a Sentinel-1 geocoded unwrapped interferogram using ISCE2's TOPS processing workflow from a set of bursts"""
 
 import argparse
 import logging
 import sys
+import warnings
 from pathlib import Path
 from shutil import copyfile, make_archive
-from typing import Optional
+from typing import Iterable, Optional
 
 import isce  # noqa
+from burst2safe.burst2safe import burst2safe
 from hyp3lib.util import string_is_true
 from isceobj.TopsProc.runMergeBursts import multilook
 from osgeo import gdal
@@ -23,15 +25,10 @@ from hyp3_isce2.burst import (
     validate_bursts,
 )
 from hyp3_isce2.dem import download_dem_for_isce2
+from hyp3_isce2.insar_tops import insar_tops_packaged
 from hyp3_isce2.logger import configure_root_logger
 from hyp3_isce2.s1_auxcal import download_aux_cal
-from hyp3_isce2.utils import (
-    image_math,
-    isce2_copy,
-    make_browse_image,
-    oldest_granule_first,
-    resample_to_radar_io,
-)
+from hyp3_isce2.utils import image_math, isce2_copy, make_browse_image, resample_to_radar_io
 from hyp3_isce2.water_mask import create_water_mask
 
 
@@ -143,7 +140,6 @@ def insar_tops_single_burst(
     bucket: Optional[str] = None,
     bucket_prefix: str = '',
 ):
-    reference, secondary = oldest_granule_first(reference, secondary)
     validate_bursts(reference, secondary)
     swath_number = int(reference[12])
     range_looks, azimuth_looks = [int(value) for value in looks.split('x')]
@@ -201,44 +197,100 @@ def insar_tops_single_burst(
         packaging.upload_product_to_s3(product_dir, output_zip, bucket, bucket_prefix)
 
 
+def insar_tops_multi_burst(
+    reference: Iterable[str],
+    secondary: Iterable[str],
+    swaths: list = [1, 2, 3],
+    looks: str = '20x4',
+    apply_water_mask=False,
+    bucket: Optional[str] = None,
+    bucket_prefix: str = '',
+):
+    validate_bursts(reference, secondary)
+    reference_safe_path = burst2safe(reference)
+    reference_safe = reference_safe_path.name.split('.')[0]
+    secondary_safe_path = burst2safe(secondary)
+    secondary_safe = secondary_safe_path.name.split('.')[0]
+
+    range_looks, azimuth_looks = [int(value) for value in looks.split('x')]
+    swaths = list(set(int(granule.split('_')[2][2]) for granule in reference))
+    polarization = reference[0].split('_')[4]
+
+    log.info('Begin ISCE2 TopsApp run')
+    insar_tops_packaged(
+        reference=reference_safe,
+        secondary=secondary_safe,
+        swaths=swaths,
+        polarization=polarization,
+        azimuth_looks=azimuth_looks,
+        range_looks=range_looks,
+        apply_water_mask=apply_water_mask,
+        bucket=bucket,
+        bucket_prefix=bucket_prefix,
+    )
+    log.info('ISCE2 TopsApp run completed successfully')
+
+
+def oldest_granule_first(g1, g2):
+    if g1[14:29] <= g2[14:29]:
+        return [g1], [g2]
+    return [g2], [g1]
+
+
 def main():
     """HyP3 entrypoint for the burst TOPS workflow"""
     parser = argparse.ArgumentParser(description=__doc__, formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-
-    parser.add_argument('--bucket', help='AWS S3 bucket HyP3 for upload the final product(s)')
-    parser.add_argument('--bucket-prefix', default='', help='Add a bucket prefix to product(s)')
+    parser.add_argument('granules', type=str.split, nargs='*', help='Reference and secondary scene names')
+    parser.add_argument('--reference', type=str.split, nargs='+', help='List of reference scenes"')
+    parser.add_argument('--secondary', type=str.split, nargs='+', help='List of secondary scenes"')
     parser.add_argument(
         '--looks', choices=['20x4', '10x2', '5x1'], default='20x4', help='Number of looks to take in range and azimuth'
     )
     parser.add_argument(
-        '--apply-water-mask',
-        type=string_is_true,
-        default=False,
-        help='Apply a water body mask before unwrapping.',
+        '--apply-water-mask', type=string_is_true, default=False, help='Apply a water body mask before unwrapping.'
     )
-    parser.add_argument(
-        'granules',
-        type=str.split,
-        nargs='+',
-        help='Reference and secondary scene names'
-    )
+    parser.add_argument('--bucket', help='AWS S3 bucket HyP3 for upload the final product(s)')
+    parser.add_argument('--bucket-prefix', default='', help='Add a bucket prefix to product(s)')
 
     args = parser.parse_args()
 
-    granules = [item for sublist in args.granules for item in sublist]
-    if len(granules) != 2:
-        parser.error('No more than two granules may be provided.')
-    if len(granules[0]) != len(granules[1]) != 1:
-        parser.error('Must include 1 reference and 1 secondary.')
+    has_granules = args.granules is not None and len(args.granules) > 0
+    has_ref_sec = args.reference is not None and args.secondary is not None
+    if has_granules and has_ref_sec:
+        parser.error('Provide either --reference and --secondary, or the positional granules argument, not both.')
+    elif not has_granules and not has_ref_sec:
+        parser.error('Either --reference and --secondary, or the positional granules argument, must be provided.')
+    elif has_granules:
+        warnings.warn(
+            'The positional argument for granules is deprecated. Please use --reference and --secondary.',
+            DeprecationWarning,
+        )
+        granules = [item for sublist in args.granules for item in sublist]
+        if len(granules) != 2:
+            parser.error('No more than two granules may be provided.')
+        reference, secondary = oldest_granule_first(granules[0], granules[1])
+    else:
+        reference = [item for sublist in args.reference for item in sublist]
+        secondary = [item for sublist in args.secondary for item in sublist]
 
     configure_root_logger()
     log.debug(' '.join(sys.argv))
 
-    insar_tops_single_burst(
-        reference=granules[0],
-        secondary=granules[1],
-        looks=args.looks,
-        apply_water_mask=args.apply_water_mask,
-        bucket=args.bucket,
-        bucket_prefix=args.bucket_prefix,
-    )
+    if len(reference) == 1:
+        insar_tops_single_burst(
+            reference=reference[0],
+            secondary=secondary[0],
+            looks=args.looks,
+            apply_water_mask=args.apply_water_mask,
+            bucket=args.bucket,
+            bucket_prefix=args.bucket_prefix,
+        )
+    else:
+        insar_tops_multi_burst(
+            reference=reference,
+            secondary=secondary,
+            looks=args.looks,
+            apply_water_mask=args.apply_water_mask,
+            bucket=args.bucket,
+            bucket_prefix=args.bucket_prefix,
+        )